(PC-21674) feat(TrustedDevice): create trustedDevice folder in src/features and move useDeviceInfo inside
