+++ src/features/trustedDevice/helpers/useDeviceInfo.ts
